LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_setpointercapture_relatedtarget-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events Capture Test - capture and relatedTarget

Test Description: This test checks if setCapture/releaseCapture functions works properly. Complete the following actions:
Put your mouse over the purple rectangle. pointerover should be received for the purple rectangle
Press and hold left mouse button over "Set Capture" button
Move your mouse. pointerover should be received for the black rectangle
Release left mouse button to complete the test.

Test passes if the proper behavior of the events is observed.



Pointer Events Capture Test

Test complete: Scroll to Summary to view Pass/Fail Results.

The following pointer types were detected: (none).

The following events were logged: (none).

Refresh the page to run the tests again with a different pointer type.


Harness Error (FAIL), message = Unhandled rejection: unimplemented
